CharacterClient/src/requirements.txt
fastapi
uvicorn
transformers
piper-tts
xttsv2
gtts
requests
TTS
huggingface_hub
